requirement.txt
azureml-sdk
azureml-defaults
pandas
sklearn